+++ .eslintrc.js
@@ -47,10 +47,8 @@ module.exports = {
 			rules: {
 				'n8n-nodes-base/node-execute-block-missing-continue-on-fail': 'off',
 				'n8n-nodes-base/node-resource-description-filename-against-convention': 'off',
-				'n8n-nodes-base/node-param-fixed-collection-type-unsorted-items': 'off',
-				'n8n-nodes-base/node-param-options-type-unsorted-items': 'off',
 				'n8n-nodes-base/node-class-description-inputs-wrong-regular-node': 'off',
-				'n8n-nodes-base/node-class-description-outputs-wrong': 'error',
+				'n8n-nodes-base/node-class-description-outputs-wrong': 'off',
 			},
 		},
 	],